Finish `daft up`
