+++ src/aws.rs
@@ -3,7 +3,35 @@ use aws_sdk_ec2::{
     operation::describe_instances::DescribeInstancesOutput, Client,
 };
 
-use crate::config::processed::AwsCluster;
+use crate::{
+    config::processed::{AwsCluster, ProcessedConfig},
+    widgets::Spinner,
+};
+
+async fn is_authenticated_with_aws() -> bool {
+    let sdk_config =
+        aws_config::defaults(aws_config::BehaviorVersion::latest())
+            .region(
+                aws_config::meta::region::RegionProviderChain::default_provider(
+                ),
+            )
+            .load()
+            .await;
+    let client = aws_sdk_sts::Client::new(&sdk_config);
+    client.get_caller_identity().send().await.is_ok()
+}
+
+pub async fn assert_is_authenticated_with_aws() -> anyhow::Result<()> {
+    let spinner = Spinner::new("Authenticating with AWS");
+    let is_authenticated = is_authenticated_with_aws().await;
+    if is_authenticated {
+        spinner.success();
+        Ok(())
+    } else {
+        spinner.fail();
+        anyhow::bail!("You are not signed in to AWS; please sign in first")
+    }
+}
 
 pub async fn list_instances(
     aws_cluster: &AwsCluster,
@@ -44,16 +72,23 @@ pub async fn list_instance_names(
     Ok(instance_names)
 }
 
-// #[cfg(test)]
-// #[tokio::test]
-// async fn test() {
-//     let aws_cluster = AwsCluster {
-//         region: "us-west-2".into(),
-//         ssh_user: "".into(),
-//         ssh_private_key: None,
-//         image_id: "".into(),
-//         instance_type: "".into(),
-//         iam_instance_profile_arn: None,
-//     };
-//     list_instances(&aws_cluster).await.unwrap();
-// }
+pub async fn instance_name_already_exists(
+    processed_config: &ProcessedConfig,
+    aws_cluster: &AwsCluster,
+) -> anyhow::Result<bool> {
+    let instance_names = list_instance_names(aws_cluster).await?;
+    let is_contained = instance_names
+        .contains(&format!("ray-{}-head", processed_config.package.name));
+    Ok(is_contained)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    #[ignore = "Depends upon global state of AWS authentication; if you want to run this test, run it manually after you've authenticated with AWS"]
+    async fn test_is_authenticated_with_aws() {
+        assert!(is_authenticated_with_aws().await);
+    }
+}
